feat: Add animating differences feature when updated
- Change delegate sturctures
